requirements.txt
pinecone-client
PyPDF2
python-dotenv
pandas
langchain
langchain_openai
langchain_community
pinecone_text
openai
langchain-teddynote  # LangChain Teddynote 확장
PyMuPDF # PDF 읽기 및 처리 (fitz 포함)
langchain-upstage  # Upstage Embeddings 모듈
